Fix case where full build is not triggered for missing _Organized folder
This bug was introduced with the following commit: https://github.com/prenetic/_arcade-organizer/commit/d58fb8fee9a31eceb334eb581ba932729d3db234

Since no "missing" checks are performed for organize folders other than A-Z, if you disable the A-Z folders and deleted the entire _Organized folder it would not be rebuilt, or only incrementally built if updates were available.

Reverting previous change to the missing folder check and instead of checking for the existence of A-Z folders, now check for the existence of the root _Organized folder to trigger a full build. This behavior matches the note already displayed to the user about forcing a full build.